### 2025-08-04 #8725 @editorialbot
[PRE REVIEW]:  WorMe: A user-oriented software for Caenorhabditis elegans length determination
**Submitting author:** @alaromaine (<a href="http://orcid.org/0000-0002-4764-0780">Anna Laromaine</a>)
**Repository:** https://github.com/group-nn-at-icmab-csic/WorMe
**Branch with paper.md** (empty if default branch):
**Version:** v2.18
