### 2025-08-12 #8725 @josepACTG
Hello, My name is Josep,
I developed the code part of WorMe length determination.
I would like to show my disposal for helping in the review of the program, and for to resolve any doubt that may arise.
Thank you
