### 2025-09-17 #8725 @alaromaine
-
- Hi @adamltyson,
- the error I guess is that list some specific figures or images from the doi, if we remove this part, they are ok, so they should be changed to
-  10.1186/1471-213X-8-28/FIGURES/2 is INVALID ==>10.1186/1471-213X-8-28
- 10.1021/ACSNANO.3C05245/ASSET/IMAGES/LARGE/NN3C05245_0012.JPEG is INVALID ==>10.1021/ACSNANO.3C05245
- 10.1021/ES5056462/SUPPL_FILE/ES5056462_SI_001.PDF is INVALID  ==> 10.1021/ES5056462/
