+++ paper/paper.bib
@@ -87,7 +87,7 @@
 @article{Tain2008,
    abstract = {Background. Dietary restriction (DR) results in increased longevity, reduced fecundity and reduced growth in many organisms. Though many studies have examined the effects of DR on longevity and fecundity, few have investigated the effects on growth. Results. Here we use Caenorhabditis elegans to determine the mechanisms that regulate growth under DR. We show that rather than a reduction in cell number, decreased growth in wild type C. elegans under DR is correlated with lower levels of hypodermal polyploidization. We also show that mutants lacking wild type sensory ciliated neurons are small, exhibit hypo-polyploidization and more importantly, when grown under DR, reduce their levels of endoreduplication to a lesser extent than wild type, suggesting that these neurons are required for the regulation of hypodermal polyploidization in response to DR. Similarly, we also show that the cGMP-dependent protein kinase EGL-4 and the SMA/MAB signalling pathway regulate polyploidization under DR. Conclusion. We show C. elegans is capable of actively responding to food levels to regulate adult ploidy. We suggest this response is dependent on the SMA/MAB signalling pathway. © 2008 Tain et al; licensee BioMed Central Ltd.},
    author = {Luke S. Tain and Encarnación Lozano and Alberto G. Sáez and Armand M. Leroi},
-   doi = {10.1186/1471-213X-8-28/FIGURES/2},
+   doi = {10.1186/1471-213X-8-28},
    issn = {1471213X},
    issue = {1},
    journal = {BMC Developmental Biology},
@@ -137,7 +137,7 @@
 @article{Srinivasan2023,
    abstract = {Experimental studies and clinical trials of nanoparticles for treating diseases are increasing continuously. However, the reach to the market does not correlate with these efforts due to the enormous cost, several years of development, and off-target effects like cardiotoxicity. Multicellular organisms such as the Caenorhabditis elegans (C. elegans) can bridge the gap between in vitro and vertebrate testing as they can provide extensive information on systemic toxicity and specific harmful effects through facile experimentation following 3R EU directives on animal use. Since the nematodes’ pharynx shares similarities with the human heart, we assessed the general and pharyngeal effects of drugs and polypyrrole nanoparticles (Ppy NPs) using C. elegans. The evaluation of FDA-approved drugs, such as Propranolol and Racepinephrine reproduced the arrhythmic behavior reported in humans and supported the use of this small animal model. Consequently, Ppy NPs were evaluated due to their research interest in cardiac arrhythmia treatments. The NPs’ biocompatibility was confirmed by assessing survival, growth and development, reproduction, and transgenerational toxicity in C. elegans. Interestingly, the NPs increased the pharyngeal pumping rate of C. elegans in two slow-pumping mutant strains, JD21 and DA464. Moreover, the NPs increased the pumping rate over time, which sustained up to a day post-excretion. By measuring pharyngeal calcium levels, we found that the impact of Ppy NPs on the pumping rate could be mediated through calcium signaling. Thus, evaluating arrhythmic effects in C. elegans offers a simple system to test drugs and nanoparticles, as elucidated through Ppy NPs.},
    author = {Sumithra Yasaswini Srinivasan and Pilar Alvarez Illera and Dmytro Kukhtar and Núria Benseny-Cases and Julián Cerón and Javier Álvarez and Rosalba I. Fonteriz and Mayte Montero and Anna Laromaine},
-   doi = {10.1021/ACSNANO.3C05245/ASSET/IMAGES/LARGE/NN3C05245_0012.JPEG},
+   doi = {10.1021/ACSNANO.3C05245},
    issn = {1936086X},
    issue = {17},
    journal = {ACS Nano},
@@ -154,7 +154,7 @@
 @article{Jung2015,
[…]
