### 2025-09-17 #8725 @josepACTG
Changed the DOIs in the paper (WorMe/paper/paper.bib).
