### 2025-09-18 #8725 @adamltyson
Thanks @alaromaine, @josepACTG, I will start inviting reviewers.

There is still one invalid DOI, would you mind taking a look? Thanks!
