### 2025-09-18 55f9f36
Update paper.bib
