+++ paper/paper.bib
@@ -154,7 +154,7 @@
 @article{Jung2015,
    abstract = {The booming nanotechnology industry has raised public concerns about the environmental health and safety impact of engineered nanomaterials (ENMs). High-throughput assays are needed to obtain toxicity data for the rapidly increasing number of ENMs. Here we present a suite of high-throughput methods to study nanotoxicity in intact animals using Caenorhabditis elegans as a model. At the population level, our system measures food consumption of thousands of animals to evaluate population fitness. At the organism level, our automated system analyzes hundreds of individual animals for body length, locomotion speed, and lifespan. To demonstrate the utility of our system, we applied this technology to test the toxicity of 20 nanomaterials at four concentrations. Only fullerene nanoparticles (nC60), fullerol, TiO2, and CeO2 showed little or no toxicity. Various degrees of toxicity were detected from different forms of carbon nanotubes, graphene, carbon black, Ag, and fumed SiO2 nanoparticles. Aminofullerene and ultraviolet-irradiated nC60 also showed small but significant toxicity. We further investigated the effects of nanomaterial size, shape, surface chemistry, and exposure conditions on toxicity. Our data are publicly available at the open-access nanotoxicity database www.QuantWorm.org/nano.},
    author = {Sang Kyu Jung and Xiaolei Qu and Boanerges Aleman-Meza and Tianxiao Wang and Celeste Riepe and Zheng Liu and Qilin Li and Weiwei Zhong},
-   doi = {10.1021/ES5056462/},
+   doi = {10.1021/ES5056462},
    issn = {15205851},
    issue = {4},
    journal = {Environmental Science and Technology},
