### 2025-09-18 #8725 @josepACTG
Hello, invalid DOI is now fixed.
