### 2025-10-13 group-nn-at-icmab-csic/WorMe#11 @Pablo1990
[minor] Confusing source code directory
Hi,

Reviewer for https://github.com/openjournals/joss-reviews/issues/9221 here. I found a confusing the name of the folder to host the code at folder 'Length_determination', I don't know if that is required in any way in MATLAB, but I'd suggest something more standard like 'source' or 'src' instead (even 'code' is better to me). In this way, people can find the code easier.

Thanks,
Pablo
