### 2025-10-13 group-nn-at-icmab-csic/WorMe#12 @Pablo1990
[major] Installation section
Referring to https://github.com/openjournals/joss-reviews/issues/9221

I think the installation [section](https://github.com/group-nn-at-icmab-csic/WorMe?tab=readme-ov-file#installation) needs major improvements:

- The link to Installation guide is broken: https://github.com/group-nn-at-icmab-csic/WorMe/blob/main/Installation/Readme.md
- The guide located in the readme is not clear at all. If you think it is best to use the binaries directly (which I feel it is), please do link it. In addition, you can either: 1. Refer to more details. 2. Clearly state that you prefer the other way, but you can also go the route of compile and create the binaries or run the code via Matlab.

I would suggest to have a simple, but effective installation guide at the READ.ME and a longer version with all the details somewhere else with possible or known issues.
