### 2025-10-13 group-nn-at-icmab-csic/WorMe#13 @Pablo1990
[major] Binaries don't work on Mac OS X intel
Referred to: https://github.com/openjournals/joss-reviews/issues/9221

I have tried both [WorMe_v2_18_iOS_local_inst.rar](https://github.com/group-nn-at-icmab-csic/WorMe/releases/download/v2.18/WorMe_v2_18_iOS_local_inst.rar) and [WorMe_v2_18_iOS_inst.rar](https://github.com/group-nn-at-icmab-csic/WorMe/releases/download/v2.18/WorMe_v2_18_iOS_inst.rar). Neither work for me out of the box. It says "The application can't be opened".

My specs: MacBook pro, macOS 15.7.1, Intel core i7.
