+++ Length_determination/WM_length_determination.m
@@ -1,184 +1,0 @@
-% WorMe_length_determination_v2_18
-disp("WorMe_length_determination_v2_18")
-
-% Developed in MATLAB 2021b version
-
-% Version 2.18
-
-% _____ PRINCIPAL VARIABLES ____
-% userSavedDocuments - directory of the program (in MATALB o compiled)
-%                   ex (using MATLAB):  'C:/Users/jllobet/Desktop/Length determination v2_16'
-%                   ex (compiled in Windows): 'C:/Users/Josep TOSHIBA/Documents/WorMe_Length_Results'
-%
-% carpeta_output - directory of res
-% ults
-%                   ex (using MATLAB):  "C:/Users/jllobet/Desktop/Length determination v2_15/Results_out/20230124_1625_C_48"
-%                   ex (compiled in Windows): 'C:/Users/Josep TOSHIBA/Documents/WorMe_Length_Results Results_out/20230407_1813_control'
-%
-% theFiles : Data structure of the images (name, folder, date, ...). Utility for obtaining images.
-
-
-
-% ------------------------------------------------
-% ______________START OF THE PROGRAM______________
-% ------------------------------------------------
-
-
-% __Determination of paths and folders__
-
-% _Path determination_
-[userSavedDocuments] = path_determination(); % local function
-
-% _Initial folder creation_
-initial_folder_creation(userSavedDocuments)
-
-% _Botation of the initial image folder_
-[theFiles, carpeta_input, nom_carpeta_input, porta_f] = obtain_theFiles_img(userSavedDocuments);
-
-
-% If the selection is correct:
-if porta_f
-    
-    % _Creation of results folder_
-    % Create the results folder.
-    [carpeta_output] = results_folder_create(userSavedDocuments, nom_carpeta_input);   
-    
-    % ___Scale determination (App)___
-    [escala_imatge] = scale_determination_main(userSavedDocuments, theFiles, carpeta_output);
-    
-    % ___Processment of the image (App)___ 
-    waitfor(app_image_processment(theFiles, carpeta_output))
-    
-    % ___Worm selection (App)___
-    waitfor(app_selection_pannel(theFiles, carpeta_output, escala_imatge))
-
-end
-
-
-% -------------------------------------------------
-% ______________FINAL OF THE PROGRAM ______________
-% -------------------------------------------------
-
-
-
-
-
-
-
-
-
-
-
-
-
-% __Local functions__
-
-
-% Note: Not to move the function path_determination from local main script.
-
-function [userSavedDocuments] = path_determination()
-% Automatically determines the path.
-%
-% Note: This function requires being in the main script.
-
-% START OF THE FUNCTION
-    
-    
-    % ___PATH DETERMINATION___
-    
-    % -_modificacio_executable: The entire try and catch of addpath is inhibited (%).
-    
-    % %_Folders_%  
-    % userSavedDocuments :  
-    %           e.g. execution in MATLAB: 'C:/Users/jllobet/Desktop/Length determination v2_12new'
-    %           e.g. execution in .exe: C/:Users/jllobet/Documents/WorMe_Length_Results
-    % currentDir         : 
-    %           e.g. execution in MATLAB: 'C:/Users/jllobet/Desktop/Length determination v2_12new'
-    %           e.g. execution in .exe: 'C:/Users/jllobet/Desktop/Length determination v2_12new/Compilacio/WM_length_determinationv212/for_testing'
-
-    
-    
-    % If execute the software instal·led (from .exe, from compiled file)
-    if isdeployed % Stand-alone mode.(Runtime)
-        [~, result] = system('path');
-        currentDir = char(regexpi(result, 'Path=(.*?);', 'tokens', 'once'));
-        
[…]
