### 2025-10-14 group-nn-at-icmab-csic/WorMe#14 @Pablo1990
[minor] Bug on adding line to measure worm
Referred to https://github.com/openjournals/joss-reviews/issues/9221

There are a few error like this one when running the code. Particularly this one happened to me when a worm was not recognised in one of your sample images:

I, then, tried to add a line. I mistakenly put a line where the worm wasn't and I had to go back and do it again pressing 'undo' first.  I got the following error the next two attempts.

```
Error using string
No constructor 'string' with matching signature found.

Error in app_interf_table_modif_drawline (line 59)
        BB_string = strjoin(string(BB.BoundingBox), ";");

Error in app_interf_boto_drawline (line 12)
    app_interf_table_modif_drawline(app)

Error in app_selection_pannel/DrawLineButtonPushed (line 447)
            app_interf_boto_drawline(app)

Error in appdesigner.internal.service.AppManagementService/executeCallback (line 171)
                callback(appOrUserComponent, event);

Error in matlab.apps.AppBase>@(source,event)executeCallback(ams,app,callback,requiresEventData,event) (line 62)
            newCallback = @(source, event)executeCallback(ams, ...

Error using matlab.ui.control.internal.controller.ComponentController/executeUserCallback
Error while evaluating Button PrivateButtonPushedFcn.
```

And, the third time, it seemed to work, annoyingly (because I don't know what I did differently).

There are other bugs when I go back and forth undoing things. However, I found that particular window very satisfying. In 1 minute, it is very easy and intuitive to go through different images and pick the ones that are correct and the ones that are not. Well done!
