### 2025-10-15 group-nn-at-icmab-csic/WorMe#14 @josepACTG
Hello @Pablo1990 , that may happen because the buttons may have a little of delaying time, and when it is pressed more than one time sometimes the functions are executed simultaneously with different executions.
Let me see on next days about this error.
Thank you for the reply.
About the image, the image processment configuration was not good enought for to select the worm.

Let me see and I am going to reply about.
