### 2025-10-16 group-nn-at-icmab-csic/WorMe#13 @Pablo1990
@josepACTG Yes, I tried the 'regular' installation via 'double-click' on the installer. It says, it can't be opened. And then I went and tried via terminal.
It might be a problem of intel and apple silicon processor, since we (annoyingly) have two different architectures in mac. I have an intel processor...
