### 2025-10-24 group-nn-at-icmab-csic/WorMe#13 @josepACTG
Hello again @Pablo1990 ,
Sorry because I am not much used to use iOS and the way programs are installed and so on.

Maybe the best option for the continuity of the program is to put apart the Linux and iOS installation.

In iOS I just tried and I have the same error of "The application cannot be oppened" despite I installed it long time ago in a iOS computer. I don't know which can be the problem.

I am asking myself if to solve the iOS and Linux problems of installations may worth. I think the best way to avoid errors in the program and to accelerate the review would be to simplify the possible OS running of the program, and just operate it in Windows. There is already the guide for Windows installation.

So, if we settled WorMe to run just in Windows the program would be more simple, with less errors, and the review of it would be closed faster and in a more direct way.
You think it is a good option?

Thank you
