### 2025-10-24 group-nn-at-icmab-csic/WorMe#12 @josepACTG
Hello @Pablo1990 , I changed the .rar to .zip format of files, and also 'repare' the borken link.
As I said in the neighbour issue, I am valuing to made just to operate WorMe in a Windows OS, avoiding iOS and Linux and avoiding so the errors on it.
Thank you!
