### 2025-10-29 8870c70
Installation guide upload. Remove iOS and Linux from documentation and paper, and installation files.
