+++ Readme.md
@@ -59,7 +59,7 @@ Please feel free to create an issue if you have any questions or comments. Devel
 
 ## Installation
 
-WorMe is a MATLAB application for Windows, macOS and Linux. It **does not require a MATLAB licence**. The software binaries and source code are found [here](https://github.com/group-nn-at-icmab-csic/WorMe/releases/latest).
+WorMe is a MATLAB application for Windows. It **does not require a MATLAB licence**. The software binaries and source code are found [here](https://github.com/group-nn-at-icmab-csic/WorMe/releases/latest).
 - Install as local software from the executable or install file.
 - Execute the source code, `WM_length_determination.m`, from the MATLAB Desktop environment software (version >2021b). (This requires a MATLAB licence).
 - Execute as a local file after installing MATLAB Runtime (version 9.11).
