+++ documentation/worme_manual.md
@@ -39,17 +39,20 @@ _Figure 1: Use of the program_
 
 ## Installation
 
-WorMe is a MATLAB based software that **does not require a MATLAB licence**. It can be easily installed and executed in one of these two options:  
-- Install as local software from the [executable file](https://github.com/group-nn-at-icmab-csic/WorMe/releases) (.exe). This does not require a MATLAB licence, because the program is compiled.
-- Execute [the source code](https://github.com/group-nn-at-icmab-csic/WorMe/source/WM_length_determination.m) from the MATLAB Desktop environment software (version >2021b). This requires a MATLAB licence but it has a slightly faster execution time.
+WorMe is a MATLAB based software that **does not require a MATLAB licence** for its use. It can be easily installed or executed from the source code.  
+See the [**Guide for install WorMe Length determination**](https://github.com/group-nn-at-icmab-csic/WorMe/Installation/README.md) for a detailed guide of installation.   
+
+In the guide, is descrived the ways WorMe can be executed:
+
+- As an [**standalone aplication**](https://github.com/group-nn-at-icmab-csic/WorMe/Installation/README.md#installation-of-worme-standalone-software). MATALB license is not required. It install easy and automatically WorMe and MATLAB Runtime 9.11 as local softwares.
+- As a [**runtime-dependent application**](https://github.com/group-nn-at-icmab-csic/WorMe/Installation/README.md#installation-of-worme-runtime-dependent-application).  MATLAB license is not required. It installs the MATLAB Runtime 9.11 and executes WorMe as a local runtime-dependent application. 
+- [**Running the code from MATLAB Desktop**](https://github.com/group-nn-at-icmab-csic/WorMe/Installation/README.md#runnign-worme-from-matlab-desktop). License is required. MATLAB version 9.11 (2021b) or higher.
+
+<br>
 
 The program requires at least 4GB of available disk space, as well as a Windows Operating System. 
 
-### Installation and use from the installer
-
-WorMe Length determination can be **easily installed from the [installer](https://github.com/group-nn-at-icmab-csic/WorMe/releases)** as a local software in Windows, macOS or Linux. **It does not require a MATLAB licence**. It can also be executed as a local file, which will install the program and does not require internet access. A [brief installation guide](../Installation/README.md) is also provided.
-
-The step-by-step installation will automatically download and install MATLAB Runtime R2021b (9.11) and WorMe Length determination as system software. [MATLAB Runtime](https://mathworks.com/products/compiler/matlab-runtime.html) is a collection of the libraries required to run MATLAB aplications, in this case WorMe Length determination.
+The use of the program from the MATLAB Desktop environment software is slightly faster, but the local-installed program is also optimal.
 
 <div align="center">
   <img src="images/use_of_the_program/WM_install.png" alt="Installation Process" width="60%">
@@ -57,17 +60,6 @@ The step-by-step installation will automatically download and install MATLAB Run
 _Figure 2: WorMe installer that triggers the installation of MATLAB runtime R2021b_
 </div>
 
-### Installation and use from the source code 
-
-The program can also be used by by running the main script [`WM_length_determination.m`](https://github.com/group-nn-at-icmab-csic/WorMe/blob/main/source/WM_length_determination.m) in MATLAB version R2021b or greater. It can be used in Windows, macOS and Linux. However, a MATLAB licence is required.  
-
-In this case, the program uses a few toolboxes which are needed for its execution (See: [How to add Add Ons in MATLAB](https://es.mathworks.com/help/matlab/matlab_env/get-add-ons.html)):
-- [Computer Vision Toolbox](https://es.mathworks.com/products/computer-vision.html)
[…]
