+++ Installation/README.md
@@ -1,18 +1,226 @@
 # Guide for install WorMe Length determination
 
-WorMe can be executed from the MATLAB Desktop by running the code (license required), or installed/executed as local by installing the MATLAB Runtime 9.11 (licesnse not required).   
+WorMe works in Windows OS. On the next, is described different ways to run WorMe and the installation guide:
+
+- As an [**standalone aplication**](#installation-of-worme-standalone-software). MATALB license is not required. It installs automatically WorMe and MATLAB Runtime 9.11 as local softwares.
+- As a [**runtime-dependent application**](#installation-of-worme-runtime-dependent-application).  MATLAB license is not required. It installs the MATLAB Runtime 9.11 and executes WorMe as a local runtime-dependent application. 
+- [**Running the code from MATLAB Desktop**](#runnign-worme-from-matlab-desktop). License is required. MATLAB version 9.11 (2021b) or higher.
+
+
+For to see the updated versions of the WorMe program, see: [**latest releases**](https://github.com/group-nn-at-icmab-csic/WorMe/releases/latest).
+
+
+
+## Installation of WorMe standalone software
+
+<div align="center">
+
+<a href="https://raw.githubusercontent.com/group-nn-at-icmab-csic/WorMe/main/Installation/Windows/Installer/WorMe_v2_18_Win_inst.exe" style="font-size:18px; font-weight:bold;">
+ <b>WorMe installer</b>
+</a>
+
+
+</div>
+
+The next steps describes the installation of WorMe as a local software. This is a easy installation process and doesn't need MATLAB license. WorMe and MATLAB Runtime is going to be installed in Windows OS.
+
+<div align = "center">
+<img src="images/WM_ld_Installation_guide/8.png" alt="WorMe Logo" width="60">
+</div>
+
+### Installation guide
+Download and install the [**executable installation file**](https://raw.githubusercontent.com/group-nn-at-icmab-csic/WorMe/main/Installation/Windows/Installer/WorMe_v2_18_Win_inst.exe). 
+
+<div align = "center">
+<img src="images/WM_ld_Installation_guide/0.png" alt="WorMe Logo" width="60">
+<br>
+↓
+<br>
+<img src="images/WM_ld_Installation_guide/1.png" alt="WorMe Logo" >
+
+↓
+<br>
+Describe the installation folder:
+<br>
+<img src="images/WM_ld_Installation_guide/2.png" alt="WorMe Logo" >
+Note: is recommended to add a shortcut to the desktop.
+
+↓
+<br>
+MATLAB Runtime installation:
+<br>
+<img src="images/WM_ld_Installation_guide/3.png" alt="WorMe Logo" >
+</div>
+
+
+Note: MATLAB Runtime is the software which allow to execute the WorMe program. It is not the main MATLAB software, and you will not need a MATLAB license. MATLAB Runtime is just a software that allow to execute the MATLAB compiled programs, like the WorMe program.
+
+
+
+<div align = "center">
+<img src="images/WM_ld_Installation_guide/4.png" alt="WorMe Logo" >
+<br>
+↓
+
+<img src="images/WM_ld_Installation_guide/5.png" alt="WorMe Logo" >
+<br>
+↓
+
+
+<img src="images/WM_ld_Installation_guide/6.jpg" alt="WorMe Logo" >
+<br>
+↓
+
+
+<img src="images/WM_ld_Installation_guide/7.jpg" alt="WorMe Logo" >
+<br>
+
+</div>
+
+Now you have WorMe Length determination installed in your computer. It is going to run by MATLAB Runtime.
+
+
+<div align = "center">
+<img src="images/WM_ld_Installation_guide/8.png" alt="WorMe Logo" >
+</div>
+
+
+For uninstall the program, you just need to go to 'Add or remove programs' of Windows.
+
[…]
