### 2025-10-29 d164a85
detail in image width
