### 2025-10-29 group-nn-at-icmab-csic/WorMe#12 @josepACTG
Hello @Pablo1990 ,
I redesigned the [Readme.md of the Installation guide](https://github.com/group-nn-at-icmab-csic/WorMe/blob/main/Installation/README.md).
I detailed every of the possibilities on the WorMe installation of the guide, and I transport it to the paper, main page Readme.md, and in the guide of the program.

I delete the iOS and Linux installation files and all the related installation procedures or comments, from all the files of the program.

Tell me if as it is right now is proper.

Thank you!
