### 2025-10-30 group-nn-at-icmab-csic/WorMe#12 @josepACTG
Thankyou @Pablo1990 , I tested WorMe in Linux and iOS on MATLAB Desktop (license required) time ago. It was working good, so I will add it in the paper.
