### 2025-10-30 031310a
emphatise about Windows, Linux and macOS use of WorMe.
