+++ paper/paper.md
@@ -46,7 +46,7 @@ For higher quality images, researchers use compound microscopy, where usually on
 In this work we present WorMe, a *Caenorhabditis elegans* length determination software. WorMe is a MATLAB Runtime software that automates the nematode's length measurements from compound microscopy images. It is open-source and user-friendly, since it works from a graphical user interface. WorMe is also versatile, because it has a wide range of settings to process many kinds of images, and it ensures data reliability since the user selects the worms to be analysed. It is also fast, as the process is computationally optimized.
 
 # Brief Description of the Program Use and Features
-WorMe is a software program implemented in *MATLAB version 9.11 (R2021b)* [@matlab], using the *Image Processing Toolbox* [@matlabimages], *Computer Vision Toolbox* [@matlabvision], *Image Acquisition Toolbox* [@matlabimaq], and *Statistics and Machine Learning Toolbox* [@matlabstats]. WorMe can be used on Windows Operating Systems, running the code from the main script `WM_length_determination.m` from the MATLAB Desktop interface. The program is also deployed as an executable using MATLAB Runtime, so it can be installed and run without a MATLAB license.
+WorMe is a software developed in *MATLAB version 9.11 (R2021b)* [@matlab], using the *Image Processing Toolbox* [@matlabimages], *Computer Vision Toolbox* [@matlabvision], *Image Acquisition Toolbox* [@matlabimaq], and *Statistics and Machine Learning Toolbox* [@matlabstats]. The program is deployed as an executable using MATLAB Runtime, so it can be installed and run without a MATLAB license on Windows Operating System. WorMe can be also used on Windows, Linux or macOS, by running the code from the main script `WM_length_determination.m` from the MATLAB Desktop interface, version R2021b or greater. 
 
 When the program is started, it prompts the user to provide the images for analysis, where one or multiple images can be selected. The scale is then set by selecting the scale bar or writing the scale value in pixels per unit.
 
