+++ source/WM_length_determination.m
@@ -1,7 +1,7 @@
 % WorMe_length_determination_v2_18
 disp("WorMe_length_determination_v2_18")
 
-% Developed in MATLAB 2021b version
+% Developed in MATLAB version R2021b
 
 % Version 2.18
 
