### 2025-10-30 #8725 @josepACTG
Hi @adamltyson ,
As @Pablo1990 point us, we couldn't execute the runtime-dependent programs of WorMe in Linux and macOS.
This made WorMe operating as free of MATLAB license as runtime-dependent in Windows operating system. It can be run in Windows, Linux and macOS in the MATLAB Desktop software which require license.
We settled the paper, documentation and the installation guide about it.
I just wanted to point out this change in the operating the program has actually.

Thank you
