### 2025-11-14 group-nn-at-icmab-csic/WorMe#15 @adamltyson
Add a link to the documentation into the repository "About" sidebar
Currently users have to scan the readme to find a link to the docs. It would be useful for this to be in the "About" section on the right-hand side (e.g. [the fiji repo](https://github.com/fiji/fiji)).

Related to, but not necessary for https://github.com/openjournals/joss-reviews/issues/9221
