### 2025-11-17 399115f
Fix formatting of user manual link in Readme.md
