### 2025-11-17 23e141c
Update README to specify software context
Clarified the purpose of WorMe in the README.
