### 2025-11-17 group-nn-at-icmab-csic/WorMe#15 @josepACTG
Hi @adamltyson , Good Idea, thank you for point it out..!
I added the link to the docs in the About section.
