+++ paper/paper.md
@@ -66,7 +66,7 @@ The user manual provides a more detailed explanation and examples of use. This d
 
 We acknowledge discussions with the ICMAB-CSIC Nanoparticles and Nanocomposites group working with *C. elegans*, Dr. Jordi Faraudo (ICMAB) for his help in the GitHub repository and M.D. Míriam Vidal for the initial tests of the software.
 
-The work has been supported partially with RTI2018-096273-B100, PID2021-122645OB-100  funded by MCIN/AEI//10.13039/501100011033/FEDER “Una manera de hacer Europa”;  the “Severo Ochoa” Programme for Centres of Excellence, in R&D  CEX2019-000917-S (FUNFUTURE), CEX2023-001263-S (MATRANS42); the Generalitat de Catalunya (2021SGR00446 Grant) and European Union’s Horizon Europe research and innovation program under grant agreement No 101057527 (NextGEM). JLL acknowledges ayudas JAE ICU, Consejo Superior de Investigaciones Científicas (CSIC), JAEICU-21-ICMAB01 (2021) in the framework of the Master of Bioinformatics and Biostatistics  (UOC, 2021),  NGA acknowledges ayudas JAEIntro, Consejo Superior de Investigaciones Científicas (CSIC),  JAEINT-24-01745, (2024), AMJ acknowledges the Ph.D. scholarship (FPU18/05190).
+The work has been supported partially with RTI2018-096273-B100, PID2024-1576370B-I00 (BACTIVE), PDC2023-145826-I00 (BIOCCHIP),  funded by MCIN/AEI//10.13039/501100011033/FEDER “Una manera de hacer Europa”;  the “Severo Ochoa” Programme for Centres of Excellence, in R&D  CEX2019-000917-S (FUNFUTURE), CEX2023-001263-S (MATRANS42); the Generalitat de Catalunya (2021SGR00446 Grant),  and the European Union’s Horizon Europe research and innovation program under grant agreement No 101057527 (NextGEM). AL participates in CSIC-Conexion Nanomedicina del CSIC, EPNOE network, and Red Nanocare 2.0, Grant RED2022-134560-T funded by MCIN/AEI/10.13039/501100011033. JLL acknowledges ayudas JAE ICU, Consejo Superior de Investigaciones Científicas (CSIC), JAEICU-21-ICMAB01 (2021) in the framework of the Master of Bioinformatics and Biostatistics  (UOC, 2021),  NGA acknowledges ayudas JAEIntro, Consejo Superior de Investigaciones Científicas (CSIC),  JAEINT-24-01745, (2024), AMJ acknowledges the Ph.D. scholarship (FPU18/05190).
 
 
 # References
