### 2025-11-18 f4df033
Update paper.md
