+++ paper/paper.md
@@ -30,23 +30,23 @@ bibliography: paper.bib
 
 # Summary
 
-*Caenorhabditis elegans* is a widely used animal model for biomedical research. The length of this nematode is commonly used to assess its development and its health. Usually, measurements are performed manually, although software that automates this process has been developed. However, most of these tools are designed for stereomicroscope images, and none work reliably for compound microscopy images. Here we present WorMe, a user-friendly software that measures *C. elegans* length in compound microscopy images. The program is versatile in handling various image types, and can analyse multiple images collectively. Furthermore, the user can accept or discard the detected objects, separate joined worms, correct erroneous measurements, and manually add new worms. The length results are presented in a spreadsheet file, with each measurement linked to its image. Graphic data can also be exported.
+*Caenorhabditis elegans* is a widely used animal model for biomedical research. The length of this nematode is commonly used to assess its development and its health. Typically, measurements are performed manually; however, software has been developed to automate this process. However, most of these tools are designed for stereomicroscope images, and none work reliably for compound microscopy images. Here we present WorMe, a user-friendly software that measures *C. elegans* length in compound microscopy images. The program is versatile in handling various image types and can analyse multiple images collectively. Furthermore, the user can accept or discard the detected objects, separate joined worms, correct erroneous measurements, and manually add new worms. The length results are presented in a spreadsheet file, with each measurement linked to its image. Graphic data can also be exported.
 
 # Statement of Need
 *Caenorhabditis elegans* is a transparent nematode widely used as a robust early-stage research and toxicology studies model. Its small size (~1mm), short life cycle (~3days), and its proliferative cycle allow for cost-effective and high-throughput experiments [@Kaletta2006]. In addition, it is estimated that 60% of human genes have a homolog in *C. elegans* [@Markaki2020].
 
 Body size is a crucial endpoint used to assess the nematode’s development that can be affected by dietary changes or altered temperatures [@Tain2008; @So2011; @Muoz-Juan2024]. Furthermore, in toxicity studies, the nematode’s body length is used to assess whether exposure inhibits growth [@Schrter2024; @Jung2015; @Srinivasan2023].
 
-The length of *C. elegans* is usually measured manually from microscopy images using FIJI-ImageJ, an image analysis software [@Schindelin2012]. This method presents several disadvantages as it is time-consuming and imprecise, since it involves manually tracing a line along the middle of each worm. Thus, the measured length would vary across attempts and experimenters.
+The length of *C. elegans* is usually measured manually from microscopy images using FIJI-ImageJ, an image analysis software [@Schindelin2012]. This method presents several disadvantages, as it is time-consuming and imprecise, since it involves manually tracing a line along the middle of each worm. Thus, the measured length would vary across attempts and experimenters.
 
-Therefore, many software solutions have been developed to automate and improve this process, such as the WormSizer plugin for FIJI [@Moore2013], WormToolBox from CellProfiler [@Whlby2012], WormLength from QuantWorm [@Jung2014], Anilength [@Jung2021] or WorMachine [@Hakim2018], among others. Most are designed to measure length from stereomicroscope images, as they capture a large number of worms in the same picture. However, the quality of these images is usually low.
[…]
