### 2025-11-18 f937592
Update paper.md
