### 2025-11-18 84ad393
Update paper.md
