### 2025-11-19 10236b2
Fix grammatical errors in paper summary
