+++ paper/paper.md
@@ -30,7 +30,7 @@ bibliography: paper.bib
 
 # Summary
 
-*Caenorhabditis elegans* is a widely used animal model for biomedical research. The length of this nematode is commonly used to assess its development and its health. Typically, measurements are performed manually; however, software has been developed to automate this process. However, most of these tools are designed for stereomicroscope images, and none work reliably for compound microscopy images. Here we present WorMe, a user-friendly software that measures *C. elegans* length in compound microscopy images. The program is versatile in handling various image types and can analyse multiple images collectively. Furthermore, the user can accept or discard the detected objects, separate joined worms, correct erroneous measurements, and manually add new worms. The length results are presented in a spreadsheet file, with each measurement linked to its image. Graphic data can also be exported.
+*Caenorhabditis elegans* is a widely used animal model for biomedical research. The length of this nematode is commonly used to assess its development and its health. Typically, measurements are performed manually; however, software has been developed to automate this process. Most of these tools are designed for stereomicroscope images, and none work reliably for compound microscopy images. Here we present WorMe, a user-friendly software that measures *C. elegans* length in compound microscopy images. The program is versatile in handling various image types and can analyse multiple images collectively. Furthermore, the user can accept or discard the detected objects, separate joined worms, correct erroneous measurements, and manually add new worms. The length results are presented in a spreadsheet file, with each measurement linked to its image. Graphic data can also be exported.
 
 # Statement of Need
 *Caenorhabditis elegans* is a transparent nematode widely used as a robust early-stage research and toxicology studies model. Its small size (~1mm), short life cycle (~3days), and its proliferative cycle allow for cost-effective and high-throughput experiments [@Kaletta2006]. In addition, it is estimated that 60% of human genes have a homolog in *C. elegans* [@Markaki2020].
