### 2025-11-20 ccea1c1
Update paper.md
