+++ paper/paper.md
@@ -20,7 +20,7 @@ authors:
     corresponding: true # (This is how to denote the corresponding author)
     affiliation: 1
 affiliations:
- - name: Institut de Ciencia de Materials de Barcelona (ICMAB-CSIC),Campus UAB Bellaterra, Barcelona, Spain
+ - name: Institut de Ciencia de Materials de Barcelona (ICMAB-CSIC), C/ dels Til.lers s/n, Campus UAB Bellaterra, Barcelona, Spain
    index: 1
  - name: "Current: #Department of Molecular Biology and Genetics. $Danish Research Institute of Translational Neuroscience—DANDRITE, Nordic-EMBL Partnership for Molecular Medicine, Aarhus University, Aarhus, Denmark"
    index: 2
