### 2025-11-27 e76fb7c
Update paper.md
