### 2025-11-27 #9221 @nuriagaralon
@adamltyson I have just changed FIJI-ImageJ to ImageJ/FIJI as you said. Thank you!
