### 2025-11-28 #9221 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
