### 2025-12-01 181eabc
Update paper.md
