### 2025-12-01 #9221 @nuriagaralon
@Kevin-Mattheus-Moerman Hello, both the software archive and the typos are now fixed, as well as some minor phrasing that was a little confusing.
Thank you!
